Fix mux bugs
Primarily this fixes an issue due to frontend/backend Receive() being
called before the channel receiver is finished using the received
message. As the Receive() doc points points out, the message is only
valid until the next call, so we need to make sure we wait before
calling Receive() again.